+++ lib/legion/extensions/swarm.rb
@@ -3,7 +3,9 @@
 require 'legion/extensions/swarm/version'
 require 'legion/extensions/swarm/helpers/charter'
 require 'legion/extensions/swarm/helpers/swarm_store'
+require 'legion/extensions/swarm/helpers/sub_agent'
 require 'legion/extensions/swarm/runners/swarm'
+require 'legion/extensions/swarm/runners/spawn_child'
 
 module Legion
   module Extensions